Allow for responsive widths
Changing width to max-width will allow the output to be resized on smaller browsers and devices. 

I teach with markdown and I need to bump up the font size quite a bit and this is causing horizontal scrolling. max-width will resize once you get smaller than 912px.

This is also how the marked app works.