In Mac OS, show previews in browser instead of HTML program
In Mac OS `open` with a path to an HTML file will invoke the default program for opening HTML files. For me and many Sublime users, this is Sublime, not Chrome.

This adds code which automatically gets the default browser on Mac OS.

Here's a Stack Overflow question I just posted, on the subject: http://stackoverflow.com/q/23506033/893113